refactor: add homeboy-release-contract; move release/deploy state types down
Prep 1/3 for the deploy+release extraction. Moves the pure release/deploy
*state* data types DOWN into a new below-core contract crate so core's
status-reporting mechanics (fleet/project/context) can reference them without
a cycle once deploy+release become their own crate.

Moved to homeboy-release-contract: DeployReason, ComponentStatus (+ its
requires_deploy accessor), ReleaseState (+ status()), ReleaseStateStatus
(+ as_str()), ReleaseStateBuckets. deploy/types.rs re-exports them so the
deploy/release code paths are unchanged; context/report.rs + fleet/status.rs
now reference the contract directly for these types.

No behavior moved (deploy::run, release::version/changelog still in core) —
that is prep 2/3 (the ReleaseProvider hook). No functional change.

cargo check --workspace --tests: GREEN.
