+++ crates/homeboy-core/Cargo.toml
@@ -27,6 +27,7 @@ homeboy-output = { version = "0.1.0", path = "../homeboy-output" }
 homeboy-paths = { version = "0.1.0", path = "../homeboy-paths" }
 homeboy-process = { version = "0.1.0", path = "../homeboy-process" }
 homeboy-product-identity = { version = "0.1.0", path = "../homeboy-product-identity" }
+homeboy-release-contract = { version = "0.1.0", path = "../homeboy-release-contract" }
 homeboy-audit-contract = { version = "0.1.0", path = "../homeboy-audit-contract" }
 homeboy-refactor-contract = { version = "0.1.0", path = "../homeboy-refactor-contract" }
 homeboy-code-audit = { version = "0.1.0", path = "../homeboy-code-audit" }
