+++ crates/homeboy-core/src/deploy/types.rs
@@ -334,46 +334,10 @@ pub struct BuildProvenance {
     pub artifact_identity: Option<ArtifactIdentity>,
 }
 
-/// Reason why a component was selected for deployment.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum DeployReason {
-    /// Component was explicitly specified by ID
-    ExplicitlySelected,
-    /// --all flag was used
-    AllSelected,
-    /// Local and remote versions differ
-    VersionMismatch,
-    /// Could not determine local version
-    UnknownLocalVersion,
-    /// Could not determine remote version (not deployed or no version file)
-    UnknownRemoteVersion,
-}
-
-/// Status indicator for component version comparison.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum ComponentStatus {
-    /// Local and remote versions match
-    UpToDate,
-    /// Local version ahead of remote (needs deploy)
-    NeedsUpdate,
-    /// Remote version ahead of local (local behind)
-    BehindRemote,
-    /// Local checkout is behind its upstream branch
-    BehindUpstream,
-    /// Remote matches a configured source checkout that is stale or detached
-    SourceStale,
-    /// Cannot determine status
-    Unknown,
-}
-
-impl ComponentStatus {
-    /// Whether deploying the configured source would advance the target.
-    pub fn requires_deploy(&self) -> bool {
-        matches!(self, Self::NeedsUpdate)
-    }
-}
+// DeployReason and ComponentStatus moved DOWN to homeboy-release-contract so
+// core's fleet/project/context status mechanics can reference them without a
+// cycle. Re-exported here so the deploy/release code keeps its paths.
+pub use homeboy_release_contract::{ComponentStatus, DeployReason};
 
 /// Compare the configured source version with the version observed on the target.
 ///
@@ -398,73 +362,9 @@ pub fn compare_deployed_versions(
     }
 }
 
-/// Release state tracking for deployment decisions.
-/// Captures git state relative to the last version tag.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ReleaseState {
-    /// Number of commits since the last version tag
-    pub commits_since_version: u32,
-    /// Number of code commits (non-docs)
-    #[serde(skip_serializing_if = "is_zero_u32")]
-    pub code_commits: u32,
-    /// Number of docs-only commits
-    #[serde(skip_serializing_if = "is_zero_u32")]
-    pub docs_only_commits: u32,
-    /// Whether there are uncommitted changes in the working directory
-    pub has_uncommitted_changes: bool,
-    /// The baseline reference (tag or commit hash) used for comparison
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub baseline_ref: Option<String>,
-    /// Warning emitted when the detected baseline may not align with the current version
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub baseline_warning: Option<String>,
-}
-
-/// High-level status derived from a component release state.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
-#[serde(rename_all = "snake_case")]
-pub enum ReleaseStateStatus {
-    Uncommitted,
-    NeedsRelease,
-    DocsOnly,
-    Clean,
-    Unknown,
-}
-
-impl ReleaseState {
-    pub fn status(&self) -> ReleaseStateStatus {
-        if self.has_uncommitted_changes {
-            ReleaseStateStatus::Uncommitted
-        } else if self.code_commits > 0 {
-            ReleaseStateStatus::NeedsRelease
-        } else if self.docs_only_commits > 0 {
-            ReleaseStateStatus::DocsOnly
-        } else {
-            ReleaseStateStatus::Clean
-        }
-    }
-}
-
-impl ReleaseStateStatus {
-    pub fn as_str(self) -> &'static str {
-        match self {
-            ReleaseStateStatus::Uncommitted => "uncommitted",
-            ReleaseStateStatus::NeedsRelease => "needs_release",
-            ReleaseStateStatus::DocsOnly => "docs_only",
-            ReleaseStateStatus::Clean => "clean",
-            ReleaseStateStatus::Unknown => "unknown",
-        }
-    }
-}
-
-#[derive(Debug, Clone, Default)]
-pub struct ReleaseStateBuckets {
-    pub ready_to_deploy: Vec<String>,
-    pub needs_release: Vec<String>,
-    pub docs_only: Vec<String>,
-    pub has_uncommitted: Vec<String>,
-    pub unknown: Vec<String>,
-}
+// ReleaseState, ReleaseStateStatus, ReleaseStateBuckets moved DOWN to
+// homeboy-release-contract (see the DeployReason/ComponentStatus note above).
+pub use homeboy_release_contract::{ReleaseState, ReleaseStateBuckets, ReleaseStateStatus};
 
 /// Result for a single component deployment.
 #[derive(Debug, Clone, Serialize, Deserialize)]
